backend/requirements.txt
# Flask and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Werkzeug==2.3.7

# Celery and Background Jobs
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
